mcap: a recording that stops mid-record keeps what it had
A robot log is very often incomplete -- the recorder was killed, the disk
filled, the battery went -- and that file is frequently the most interesting
one on the machine, because it is the run that failed. Ferroscope refused all
of it, including the messages written correctly before the failure.

`stream_recovering` keeps every whole record and REPORTS where the file stops.
128,388 of 255,364 messages out of a recording cut exactly in half. What is
lost is the records inside the unfinished one: a chunk is one record, so 64 KB
chunks lose at most the last 64 KB of messages, and a file written as a single
chunk recovers only its header.

Recovery is opt-in and announces itself, and that is the part that matters. A
short file that reads like a complete one is how a diff comes to report
divergence that is really absence. So `inspect` stays strict by default and
names `--recover` in its error; `--recover` prints a TRUNCATED banner above the
listing; `diff` has no recovering path at all.

It recovers, it does not repair. A CRC mismatch or a nonsense record length is
still an error -- those mean the bytes are not what they claim to be, which is
a different thing from a file that ends early.

Two defects found on the way, both in what the errors SAY:

The truncation offset was an index into a buffer that `compact()` resets to
zero, so a 53 MB recording cut at 26.8 MB reported "truncated at byte 0" -- a
number a reader would go and look at, pointing at the wrong place. It is now a
file position: offset + 9 + bytes-present accounts for exactly the file that is
there, checked in CI, and mutation-tested against the byte-0 regression.

`verify` answered "has no Ferroscope receipt, or its payloads are unreadable"
for every failure, because `verify_streaming` returns an Option and the reason
was already gone. A truncated file and a file in a codec this build cannot read
both came back as a missing receipt, sending the reader after the wrong thing.
It now asks the reader why, on the failure path only.

Four tests, one of which pins a boundary I got wrong first: a tail shorter than
a record header is NOT truncation, because that is exactly what a file still
being written looks like.

+++ crates/ferroscope-mcap/src/read.rs
@@ -110,7 +110,7 @@ pub fn record_spans(bytes: &[u8]) -> Result<Vec<RecordSpan>> {
     while pos < body_end {
         if body_end - pos < 9 {
             return Err(Error::Truncated {
-                offset: pos,
+                offset: pos as u64,
                 want: 9,
                 have: body_end - pos,
             });
@@ -126,7 +126,7 @@ pub fn record_spans(bytes: &[u8]) -> Result<Vec<RecordSpan>> {
             Ok(n) => n,
             Err(_) => {
                 return Err(Error::Truncated {
-                    offset: pos + 9,
+                    offset: (pos + 9) as u64,
                     want: usize::MAX,
                     have: body_end - pos - 9,
                 });
@@ -134,7 +134,7 @@ pub fn record_spans(bytes: &[u8]) -> Result<Vec<RecordSpan>> {
         };
         if body_end - pos - 9 < len {
             return Err(Error::Truncated {
-                offset: pos + 9,
+                offset: (pos + 9) as u64,
                 want: len,
                 have: body_end - pos - 9,
             });
@@ -191,7 +191,7 @@ fn read_inner(bytes: &[u8], require_end_magic: bool) -> Result<Log> {
         let len = c.u64()? as usize;
         if c.remaining() < len {
             return Err(Error::Truncated {
-                offset: c.pos,
+                offset: c.pos as u64,
                 want: len,
                 have: c.remaining(),
             });
@@ -285,7 +285,7 @@ pub(crate) fn parse_attachment(body: &[u8]) -> Result<Attachment> {
     let n = b.u64()? as usize;
     if b.remaining() < n + 4 {
         return Err(Error::Truncated {
-            offset: b.pos,
+            offset: b.pos as u64,
             want: n + 4,
             have: b.remaining(),
         });
@@ -338,7 +338,7 @@ fn read_chunk(log: &mut Log, body: &[u8]) -> Result<()> {
     let n = b.u64()? as usize;
     if b.remaining() < n {
         return Err(Error::Truncated {
-            offset: b.pos,
+            offset: b.pos as u64,
             want: n,
             have: b.remaining(),
         });
@@ -365,7 +365,7 @@ fn read_chunk(log: &mut Log, body: &[u8]) -> Result<()> {
         let len = inner.u64()? as usize;
         if inner.remaining() < len {
             return Err(Error::Truncated {
-                offset: inner.pos,
+                offset: inner.pos as u64,
                 want: len,
                 have: inner.remaining(),
             });
